Rename cache utilities and files
