+++ src/parcel_service/infrastructure/repository/__init__.py
@@ -1,4 +1,4 @@
 from .parcel import ParcelRepository
 from .outbox import OutboxEventRepository
-from .parcel_combain import ParcelCombinedRepository
+from .parcel_combine import ParcelCombinedRepository
 from .parcel_type import ParcelTypeRepository
